pyproject.toml
[project]
name = "WeatherLens"
version = "0.1.0"
description = "Collect weather data and visualize it"
readme = "README.md"
requires-python = "~=3.10"
dependencies = [
    "django>=5.1.5",
    "djongo>=1.2.31",
    "mypy>=1.11.1",
    "plotly>=5.24.1",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.6.2",
]

[tool.uv]
dev-dependencies = ["mypy==1.11.1", "ruff==0.6.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff]
fix = true
line-length = 100
target-version = "py310"

[tool.ruff.lint]
preview = true
select = ["I"]
